# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Pine Bank API
  version: 1.0.0
  description: |
    Customer and admin REST API for Pine Bank. JSON, RFC 7807 errors.
    Money-movement endpoints require an `Idempotency-Key` header.
servers:
  - url: https://api.pinebank.com/api/v1
    description: Production
  - url: https://api-staging.pinebank.com/api/v1
    description: Staging
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      schema: { type: string, maxLength: 80 }
      required: true
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: Issue access + refresh tokens
      security: []
      responses:
        '200': { description: ok }
        '401': { description: invalid credentials or MFA required }
        '429': { description: rate limited / account locked }
...
  /transfers/internal:
    post:
      summary: Internal transfer between own accounts (PIN required)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: posted
        '400':
          description: validation error or invalid PIN
        '409':
          description: insufficient funds / account inactive
  /transfers/ach:
    post:
      summary: Outbound ACH transfer
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
  /transfers/wire/domestic:
    post:
      summary: Domestic wire (Fedwire)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
...